literal and identifier -> operand; filterUtil -> util; shorter tabs; moved behaviors from index.js to Hypergrid.js
