+++ src/html/markup.html
@@ -222,7 +222,7 @@ exports.filterTrees = function() {/*
         </section>
 
         <header id="columnsSQL" class="tabz-bg2">
-            SQL syntax
+            SQL
         </header>
 
         <section class="filter-expression-syntax tabz-bg2">
@@ -237,7 +237,7 @@ exports.filterTrees = function() {/*
         </section>
 
         <header id="columnsCQL" class="tabz-bg1">
-            Column Filter syntax
+            CQL
         </header>
 
         <section class="filter-expression-syntax tabz-bg1">
@@ -317,7 +317,7 @@ exports.filterTrees = function() {/*
         </section>
 
         <header id="tableSQL" class="tabz-bg2">
-            SQL syntax
+            SQL
         </header>
 
         <section class="filter-expression-syntax tabz-bg2">
